Weekly Game Jam 39 - Take A Picture - Huge amount added and improved
The UI was improved.
The lose screen was added.
The exit/goal was added and coded with a script.
The options menu is now a how to play screen.
A bunch of other littlethings and there's more to come.
Next will be adding in the actual paintings and fixing/improving the Enemy AI.